Remove the Sponsor button.
This is part of the template I (@Julian) use, but TBD on what this
project's funding model will be, and whatever it'll be, it likely won't
be strictly "fund Julian" (though if you care to do so feel free to find
me elsewhere :).

But this project will hopefully be supported by/via/alongside broader
members of the community, so let's get rid of this on the off-chance
someone throws money at me expecting it to go specifically towards
Bowtie itself.