Invert the order of smoke testing during build.
This *has* to be wrong, but I don't understand why/how image builds
are building an image then immediately testing some previous version
of it, as if they've either pulled the old image or not persisted the
new one.

An example is:

    https://github.com/bowtie-json-schema/bowtie/actions/runs/3842547528/jobs/6543945399

where a new version was built, but then the smoke tests fails -- but
that failing line does not exist in the new image, it was removed, and
the commands work fine when run locally, which again makes me suspect
weird caching behavior on GitHub action runners.