Add a mechanism for passing through arbitrary run metadata
Right now this is only used by bowtie suite (to pass through the test
suite commit info for interlinking).

In theory we can add an argument to the CLI commands to pass through
some random JSON file too to let people pass other stuff if needed, but
for now this is the main thing we want to show up.

Closes: #26