Stress the purpose of the dialect response further in the harness tut.
ok: true / ok: false in harnesses does *not* indicate whether the
dialect is supported or not, that is already known via the start
response.

It controls implicit schemas.

This has confused a number of implementers adding harnesses, most
recently https://github.com/bowtie-json-schema/bowtie/pull/416#discussion_r1302954189

Hopefully this makes it slightly clearer what this is for.