Remove the old Jinja2-based UI, along with `bowtie report`.
The CLI command we can/should likely re-add, but on top of the UI.

In other words, perhaps running it should momentarily open a local UI,
with the results piped into a local-report.

For now though, this is highly likely to go completely stale, if it
hasn't already.

This also adds a `bowtie badges` CLI command though likely that should
be rethought at some point too.