Fix the dotnet-jsonschema-net harness.
Previously essentially every test began failing with e.g.:

    Unhandled exception. System.TypeLoadException: Could not load type 'Json.More.JsonSerializerOptionsExtensions' from assembly 'Json.More, Version=1.0.0.0, Culture=neutral, PublicKeyToken=17ed63f672b0e278

Json.More seems to be another package part of JSON Everything, and
specifically seems to be a dependency of JsonSchema.Net.

It had a recent release -- it seems odd that we now mention a transitive
dependency, but I don't really understand enough about .NET to know
what's happening here or be sure whether there's an upstream bug or not.

CC @gregsdennis though in case.