Don't send draft-next schemas to harnesses when running the suite.
draft-next support in general is #40, so this will be modified when
that's done.

But for now, some implementations, specifically AJV, will blow up if
they even *see* a schema with this $schema.

Given that in theory that's not totally wrong (implementations can
complain if they see unknown dialects), and that the only reason these
are being included is because the test suite doesn't have a firm way to
differentiate between which test registry goes with which version, it's
nicer to filter these out completely so ajv and others pass more tests.