Fix java-json-schema-validator logic bug.
Missing braces in an if condition means this was doing crazy things.

Specifically, if no registry was present, we constructed no
JsonSchemaFactory, and basically then didn't respond to a request.

This manifested most directly when running `bowtie validate`, which only
runs one case.

But the same error was present when running the suite -- the only reason
later tests succeed was that once we hit a test which *did* have a
registry we'd successfully build our factory.

The formatting here now more closely matches other harnesses in Java,
and successfully smoke tests, along with reducing the number of test
failures.

Thanks Java1?!?!?