Improve what it means for two reports to be equal.
This is tedious and only partially leverages attrs's eq (even though it
supports equality 'key's).

Really for the behavior we're after we probably should replace `_cases`
and `_results` with lists -- Seq is just an `int` and really serves just
as a matching index for getting results from implementations. I've
planned on refactoring that more heavily anyways, so it's probable that
that will change. For now the impact there is just one xfailed test,
plus the fact that the `__eq__` here does some tedious work.