Better separate the layers which do harness communication.
Introduce HarnessClient as the main way to talk to an implementation,
which is separate from container-specific nonsense. In part this will
help if/when we get to #19, as well as making testing lots easier by
allowing for fakes in a more straightforward way.

Container communication is now done at a lower layer in
bowtie._containers, meaning other things don't have to deal with
aiodocker's inadequacies.

There's still plenty to improve here, like:

    * Implementation still probably has an attribute or two that could
      move elsewhere -- but now .info() is no longer randomly accessed
      just to see whether we had successful startup -- we now check that
      before creating an Implementation instance
    * HarnessClient uses .transition() to store state that will be
      needed if we have an unexpected restart -- though it does less
      error checking than we do the first time around, so in theory
      there's room for bugs for an implementation harness which does
      something flaky while being replayed. Also there's some leakage
      there where really we have a state machine and we're trying to
      represent that. Some more refactoring will eventually help there.
    * there's tons of typing nonsense because pyright seems unable to be
      convinced that @command returns a command. As usual, I wasted
      enough precious time before just giving up.

But let's commit here, and we can do further cleanup.