Loosen the Python types for expected results to allow nonhomogeneity.
We had this before and then I tightened it -- even though it's not
really likely today that anything will pass in test cases where only
some of the tests have expected results, there's no real reason we
couldn't support such a thing, and it wasn't really being enforced
(neither in Python nor in the JSON Schema) that the array was
homogeneous, so we might as well not pretend.

We can either beef up support or else really enforce this later, if it
matters.