Sigh. Fix sending invalid registries when running draft 3/4 suites.
tree.json is not a valid Draft <6 schema, and contains no $schema, but
was being sent to implentations which may strictly reject these schemas
under those dialects and have no way to know what version they were
intended for.

This appears to be the only current schema where this is the case.

Obviously this should be fixed in the test suite but it's easier to hack
around here for now.