Double check reports are valid after we generate them.
Given the storage space woes, commands can go wrong in fun ways,
producing truncated JSON (lines).

This is one quick way to double check we've produced correct output.

I'm not 100% sure why we don't also get a non-zero exit status, but
maybe we do and we're just not seeing it when the runner fills up.

After #831 is done we should also emit this to GITHUB_STEP_SUMMARY,
since we may as well see this info within the GitHub UI too.