Add a first UI+CLI integration test for report parsing.
Clearly we need more of these, but at least we can now iterate without
being completely blind.

Also, building images from JS appears not trivial?! so we do it outside
of the test suite here, and assume podman (rather than what we do in the
backend tests, where we can rely on the Docker API to work with
whatever).

Refs: #706, though doesn't go all the way out to rendering.