Remove our temporary pull of images in the report workflow.
I somewhat suspect that the missing piece of the puzzle for why c18be04
didn't fix the issue was simply because we were using the released
version of Bowtie in workflows, and that didn't have the fix.

Now that we have a release with the fix, this hopefully will work
without doing the pull outside of Bowtie -- let's find out.