No need for the full commit SHA in run metadata.
It's just long. Use the short SHA instead, as we have the
URL.

'Interestingly' I don't see an attribute on commit objects which
has this info so it's just done via string slicing.

That *should* be incorrect I think, sometimes git decides to go
from 7 to 8 if I recall correctly, but for our use case (both
where the repos are small and also because it's just for diagnosis)
this probably isn't anything to think hard about.