Don't have lots of extra padding on mobile.
Remove the hardcoded w-75s everywhere, and use col, which
seems to do this OK.

We likely/may move to a 3 column layout, in which case just
the container-xxl alone seems like it's sufficient, but we can
tweak again at that point.

Also use <main>, since why not, it seems appropriate, though we seem
like we should be doing the top level element in some place other than
the dialect report one -- stuff looks a bit too "separated" to me.
Something to look into.