Don't pre-pull images in codespaces.
In GitPod the advantage to doing this was caching, and we had it run in
the background.

We seem to get neither when running this in Codespaces, so we may as
well defer until later -- if someone runs an image in Bowtie it will
pull it, and if they only use one, only one will be pulled, otherwise we
can spend this time waiting then.