Generate our list of implementations simply by listing the directory.
It's named just right, the directory names are our image names.

Write that out, and we include the generated file in our distributions.

I don't see a way of cleaning up via a context manager, so even though
it's probably safe to abandon the generated file, we use finalize() to
remove it.