And now actually use and test our generated data.
We use the file we'll generate at build time whenever the package
is installed.

For editable installs we do something similar to what we're doing
for Dialect.known (which is slightly clunky but works) -- namely
we just know we're sitting inside a source directory so we find
the data we need directly.

Possibly we could combine this with the build script a bit better
but this should work.