Let Dialect deal with knowing the right baseURI.
And introduce implementationMetadataURI even though we have no
Implementation class (just an interface) to put the method on yet.

Means nothing else other than these two things need to muck with paths
explicitly.

Probably there's an even better way to share these paths with the
backend but this is ok for this second.