Do random things because the JS ecosystem is nuts.
I understand none of this, but it seems like after 1c523d3 (which simply
caused a file containing import.meta to be imported in the test suite)
that probably what's happening is Jest can't cope with ESM modules,
whatever that is, and even though we do the thing that the error message
complains about (setting --module to esnext), that doesn't affect Jest,
nor does just setting preset: "ts-jest/presets/default-esm" (that would
be too easy), you instead set some random scary sounding
NODE_OPTIONS=--experimental-vm-modules and then *still* do random things
to now cope with the fact that Vite (which is what's setting
import.meta.env.MODE) is apparently not running in whatever way this is
executing itself so those are all unset, so OK set them to random decent
testing values.

Load of nonsense.

If anyone knows the correcter way to fix this at any point, please feel
free to help do so.