Introduce (internally) the notion of validator registries
Use them to enable report schema validation.

A validator registry is a referencing registry which knows how to
produce validators for JSON Schemas which live within the registry.

Ultimately this is functionality intended to upstream into referencing,
but it's useful to suss it out here first as Bowtie needs it itself.

Only one implementation in present (for jsonschema, which we already
were using here).