(Temporarily?) make container connections manage their Docker instance
This removes some global shared state which makes implementing
connectables a bit harder -- specifically if we want to share this
instance we need some way to share it across *only* Container
connectables.

Possibly that's something we can tweak, but for now given this works as
well and that we're not creating 1000 of these it seems safe to
change.