Remove the FIXME for checking out the repo.
We still checkout the repo, but now essentially to use the
up to date version of the action. We could I guess switch to doing
uses: bowtie-json-schema/bowtie@main.

But at least for implementation-listing, we do indeed now have
bowtie filter-implementations (though it's currently broken in CI
for some reason...).