Fix a race condition in the container test fixture
We were suppressing DockerErrors in the container fixture for calls to
image.delete() -- but that's not enough, as it's not guaranteed whether
the image is deleted there or by the *image* fixture (if another test
has built the same image -- after all, we scope them to module scope).

That means this can (and sometimes did) blow up with errors coming from
cleanup in the *other* image fixture where we didn't suppress them.

What's here is now less general, but we're due to refactor things here
especially now that we have connectables and may be able to reduce the
number of places we spin up containers in the tests entirely.