Add support for "direct" connectables.
This is useful for directly speaking to an implementation which we can
import.

Right now it's somewhat a POC, as we only add support for
python-jsonschema in this way (even amongst Python implementations in
theory we could add the others, not to mention have CFFI wrappers / PyO3
for C ABI and Rust implementations).

But it's something to build on -- and it will make some of the tests
faster if we refactor, as even in integration tests we don't always care
about container-specific behavior, so we can move to using these,
especially after adding direct connectables for lintsonschema and other
fake implementations.

Another good improvement would be somehow sharing the implementation
here with the python-jsonschema "real" container implementation.

Refs: #19