Slightly simplify the version-specific json_schemer implementation.
Each section fully determines the behavior for that version, and the
language more closely matches what we're using in bowtie._registry.

(More trivially, it also seems wrong to call these yoda conditions,
and maybe it's even a false positive.
The thing on the *left* should be the value you "care about", and here
we just need to get the right comparison behavior.)