Special-case fastjsonschema's image build to allow smoke failure.
It now fails our improved smoke test for a second reason, namely that it
mis-compiles `not: {$schema: ...}` (i.e. the presence of $schema in a
subschema there).

We can remove this when its upstream support improves, but the existing
PR to fix the first issue seemed to have stalled a bit, so we simply can
skip failing here and restore our "full" test.