Be a bit looser about dialect URIs with empty fragments.
As URIs, the URI http://example.com/foo# and http://example.com/foo are
*not* the same / equal / normalized to each other.

But each media type (of which JSON Schema is one) defines how to handle
this empty fragment case -- and JSON Schema says they're equivalent.

So http://json-schema.org/draft-07/schema# is the same dialect as
http://json-schema.org/draft-07/schema.

This comes up if an implementation "mis"-reports the dialect URI it
supports with-or-without the fragment (which #1397 uncovered).

We could do this more robustly, and maybe we'll tweak at some point, but
for now this at least unbreaks the build; downstream is likely to "Fix"
the issue as well.