Let direct connectables take connectable parameters.
This allows for a direct connectable which changes its behavior based on
the provided parameter(s).

Right now, this is only used to generalize the `only_draft3` miniature
so that it can claim to support any specific dialect -- in other words,
this sort of makes the tests we have nicer. So instead of:

    bowtie validate -i direct:bowtie.tests.miniatures:only_draft3

we have

    bowtie validate -i direct:bowtie.tests.miniatures:only_supports,dialect=draft3

But this functionality is both evidently generally useful and needed for
tests we'll write for other functionality related to `bowtie trend`
(see e.g. #1384).

We don't yet ensure that a specific direct connectable emits a friendly
error message if provided with a parameter that it doesn't really take.
That's a good follow-up.